bind: Automatically update zone serial.
The serial is only updated if the zone file changed.

A grain contains the current serial. If the grain is not found, the current
date is assumed (so every resulting file will be up to date).

The zone file is rendered with the old serial but the new data, to detect
changes that need a serial update. If the file changes, it will be rendered
again with the new serial. The new serial is then used to update the grain.

Overview over the steps:

- retrieve the old serial from grains
- if not found:
  - assume first revision of the current day
- render file with that old serial
- if file changed:
  - calculate new serial
  - render file again with new serial
  - store serial in grain
- else:
  - no changes, do nothing

Calculating the new serial simply means either increaing the revision or
using the new date.