postfix: Make DH parameter file handling more dynamic.
It works like this now:

If the dhparams.pem file is found on the master, it is used, otherwise a
new one is generated on the minion.

If the optional parameter master_dhparams is True, the is file never
generated, but downloaded from the master, which will fail if it does
not exist.